+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - chore: Improved description of username option
 - feat: Add --icon-url option
 - refactor: Remove `Slkecho::Options` initialize method
+- refactor: Summarize arguments of `Slkecho::SlackRequest::PostMessage#request` to params
 
 ## [v1.2.1] - 2024-01-08